requirements.txt
Flask
flask_pymongo
werkzeug
bson
face_recognition
Pillow